Task 7.1 solved
